[#751] Moved signature activity and extracted deprecated method
